[Refactor] Money 클래스 및 테스트 수정
 Money에서 정적팩토리를 추가하였다.
